create EditImageCard and EditImageCardComponent to seperate logic from ui (#295)
seperate EditImageCard into container 
add tests for editimagecard
